requirements.txt
duckdb
pyspark
pytest
pandas
numpy
streamlit
altair
fastapi
uvicorn
google-cloud-bigquery
pandas-gbq
great_expectations